basic conversion to sdl2
At this point the project compiles, but segfaults because the input lookup table does not work
anymore. A quick solution is simply migrating to SDL_SCANCODE_* instead of SDLK_, but I'm not sure
what the effect will be. Since I'm mostly interested in running this on a switch, I don't really
care as long as I'm able to test stuff on my pc for now.

Also, since this project is small, I could migrate everything off surfaces to textures, I'm not sure
if this commit breaks rendering of the game since it segfaults on the menu, guess I'll find out soon